+++ out/command-runner/results/CommandRunnerParseOnlyResult.js
@@ -16,15 +16,59 @@ class CommandRunnerParseOnlyResult extends CommandRunnerResultBase_1.CommandRunn
         this.size = undefined;
         this.size_unit = undefined;
         this.tags = [];
+        this.json_sections = [];
+        this.remote_contents = [];
         this.input_path = input_file;
         this.script_path = script_file;
         this.input_file_name = path.basename(input_file);
         this.script_file_name = path.basename(script_file);
         this.parse_info();
         this.parse_metrics();
+        this.parse_json_sections();
     }
     parse_metrics() {
         this.metrics = ResultMetric_1.ResultMetric.parse_from_text(this.raw_data_stripped);
+    }
+    parse_json_sections() {
+        let regexp_start = /(?:\r|\n){/g;
+        let match;
+        while (match = regexp_start.exec(this.raw_data_stripped)) {
+            try {
+                let result = JSON.parse(this.parse_json_section(match.index + match.length));
+                this.json_sections.push(result);
+                if (result['event-type'] === 'record-step') {
+                    let record = result['record'];
+                    if (record) {
+                        if (record.content) {
+                            this.remote_contents.push(record.content);
+                        }
+                    }
+                }
+            }
+            catch (_a) {
+            }
+        }
+        console.log(this.remote_contents);
+    }
+    parse_json_section(offset) {
+        let level = 0;
+        let content = '';
+        for (let i = offset; i < this.raw_data_stripped.length; i++) {
+            let char = this.raw_data_stripped[i];
+            switch (char) {
+                case '{':
+                    level++;
+                    break;
+                case '}':
+                    level--;
+                    break;
+            }
+            content += char;
+            if (level <= 0) {
+                break;
+            }
+        }
+        return content;
     }
     parse_info() {
         let match_running = /Running '([^']+)/g;