+++ src/command-runner/results/CommandRunnerParseOnlyResult.ts
@@ -16,6 +16,10 @@ export class CommandRunnerParseOnlyResult extends CommandRunnerResultBase {
 
     public tags : ResultMetricTag[] = [];
 
+    private json_sections : any[] = [];
+
+    public remote_contents : string[] = [];
+
     public constructor(input_file : string, script_file : string, raw_data : string) {
         super(raw_data);
         this.input_path = input_file;
@@ -24,10 +28,62 @@ export class CommandRunnerParseOnlyResult extends CommandRunnerResultBase {
         this.script_file_name = path.basename(script_file);
         this.parse_info();
         this.parse_metrics();
+        this.parse_json_sections();
     }
 
     parse_metrics() {
         this.metrics = ResultMetric.parse_from_text(this.raw_data_stripped);
+    }
+
+    parse_json_sections() {
+        let regexp_start = /(?:\r|\n){/g;
+        let match;
+
+        while (match = regexp_start.exec(this.raw_data_stripped)) {
+            try
+            {
+                let result = JSON.parse(this.parse_json_section(match.index + match.length));
+                this.json_sections.push(result);
+                if (result['event-type'] === 'record-step') {
+                    let record = result['record'];
+                    if (record) {
+                        if (record.content) {
+                            this.remote_contents.push(record.content);
+                        }
+                    }
+                }
+            }
+            catch {
+
+            }
+        }
+
+        console.log(this.remote_contents);
+        
+
+    }
+
+    parse_json_section(offset : number) {
+        let level = 0;
+        let content = '';
+        for (let i = offset; i < this.raw_data_stripped.length; i++) {
+            let char = this.raw_data_stripped[i];
+            switch (char) {
+                case '{':
+                    level++;
+                    break;
+                case '}':
+                    level--;
+                    break;
+            }
+
+            content += char;
+            if (level <= 0) {
+                break;
+            }
+        }
+
+        return content;
     }
 
     parse_info() {