+++ src/gui/CommandRunnerResultView.ts
@@ -24,6 +24,22 @@ export class CommandRunnerResultView {
         if (this.panel === undefined) {
             this.panel = vscode.window.createWebviewPanel('commandRunnerResult', title, vscode.ViewColumn.Beside, { enableScripts: true, retainContextWhenHidden: true });
             this.panel.onDidDispose((e : void) => { this.panel = undefined; });
+            this.panel.webview.onDidReceiveMessage(message => 
+                {
+                    console.log(message);
+                    switch (message.command) {
+                        case 'show_data':
+                            if (message.data) {
+                                for (let data_string of message.data) {
+                                    vscode.workspace.openTextDocument({ content: data_string }).then((onfulfilled) =>
+                                    {
+                                        vscode.window.showTextDocument(onfulfilled);
+                                    });
+                                }
+                            }
+                            break;
+                    }
+                });
         }
 
         if (this.panel !== undefined) {